Add files via upload
This notebook contains my first machine learning homework. The first task was to compare the scikit-learn linear regression model with the model I created manually.
The second task is to create different manual linear regression models:
- Three models with the same learning rate value but different number of iterations,
- Three models with the same number of iterations but different learning rates.
The second task is to understand how the loss function works and the importance of selecting the right leaning rate and number of iteration values.