Add files via upload
Task content:
Find a set on which to prepare a regression model.
a. Prepare analysis
b. Then prepare regression models. Use all the models learned in this
class.<br>
c. Select parameters for them manually (you can test several possibilities, but do not use deafult parameters.
d. Compare the results of the models with each other and choose the best one.
e. Try to explain why this particular model is the best (the idea is to find the
the reason why model X did better than the others. I leave a lot of room for maneuver here and in
case of problems with the explanation you will have to googling and try to understand even
superficially what is going on inside the models)