Add files via upload
Consider the dataset about loans, available here:
https://www.kaggle.com/datasets/wordsforthewise/lending-club?resource=download
a. This collection contains more than 2.2million rows and 150 columns. Once loaded, limit the number of observations to a randomly
selected 20,000.
b. Create a classification model that can predict whether a person is creditworthy,
based on the input data. Exclude columns where more than 50% of observations are missing and then
consider the model based on:
1. all input data
2. 50% of randomly selected columns
3. columns selected using PCA - select the number of features so that a minimum of 50% of the variance is
explained
4. columns selected using LDA - select the number of features so that a minimum of 50% of the variance is
explained
5. columns selected using feature importance from any tree algorithm. Let the number of
columns used should be the average of the number of columns selected by PCA and LDA.
c. "Play around" with the number of inputs that are left after reduction by PCA and LDA. How does the number of variables
affect the result of the model? For each number selected, check the percentage of explained variance